+++ .tools/update_json.php
@@ -100,7 +100,7 @@
                 if (isset($iconsToRename[$icon])) {
                     $newName = $iconsToRename[$icon];
                     $dev[$devName]['configuration']['icon'] = $newName;
-                    echo "  'icon' renamed from '${icon}' to '${newName}'.\n";
+                    echo "  'icon' renamed from '{$icon}' to '{$newName}'.\n";
                     $devUpdated = true;
                 }
             }
@@ -148,7 +148,7 @@
                 // Removing some useless commands
                 $cmdsToRemove = ['SWBuildID', 'Get-SWBuildID', 'Get SWBuildID'];
                 if (in_array($cmdJName, $cmdsToRemove)) {
-                    echo "  Cmd '${cmdJName}' REMOVED.\n";
+                    echo "  Cmd '{$cmdJName}' REMOVED.\n";
                     $devUpdated = true;
                     continue;
                 }
@@ -205,7 +205,7 @@
                     $commands2[$key] = $value;
                     $commands2[$key]["use"] = $new;
                     $devUpdated = true;
-                    echo "  Cmd '${cmdFName}' renamed to '${new}'.\n";
+                    echo "  Cmd '{$cmdFName}' renamed to '{$new}'.\n";
                 }
 
                 // zbReadAttribute => act_zbReadAttribute
@@ -243,7 +243,7 @@
                         // echo "pArr2=".json_encode($pArr2)."\n";
                         $pX = $pArr2[0];
                         $vX = $pArr2[1];
-                        // echo "px=${pX}, vx=${vX}\n";
+                        // echo "px={$pX}, vx={$vX}\n";
 
                         switch ($pX) {
                         case "ep":
@@ -287,48 +287,48 @@
                     // Attribute type correction
                     if (($clustId == "0008") && ($attrId == "0000") && ($attrType != "20")) {
                         $attrType = "20";
-                        echo "  Cmd '${cmdJName}': Attribute type corrected.\n";
+                        echo "  Cmd '{$cmdJName}': Attribute type corrected.\n";
                     }
 
                     $cmd['use'] = "act_zbConfigureReporting2";
                     $params = '';
                     if ($ep !== '')
-                        $params .= "ep=${ep}";
+                        $params .= "ep={$ep}";
                     if ($clustId !== '') {
                         if ($params != '')
                             $params .= "&";
-                        $params .= "clustId=${clustId}";
+                        $params .= "clustId={$clustId}";
                     }
                     if ($attrId !== '') {
                         if ($params != '')
                             $params .= "&";
-                        $params .= "attrId=${attrId}";
+                        $params .= "attrId={$attrId}";
                     }
                     if ($attrType !== '') {
                         if ($params != '')
                             $params .= "&";
-                        $params .= "attrType=${attrType}";
+                        $params .= "attrType={$attrType}";
                     }
                     if ($minInterval !== '') {
                         if ($params != '')
                             $params .= "&";
-                        $params .= "minInterval=${minInterval}";
+                        $params .= "minInterval={$minInterval}";
                     }
                     if ($maxInterval !== '') {
                         if ($params != '')
                             $params .= "&";
-                        $params .= "maxInterval=${maxInterval}";
+                        $params .= "maxInterval={$maxInterval}";
                     }
                     if ($changeVal !== '') {
                         if ($params != '')
                             $params .= "&";
-                        $params .= "changeVal=${changeVal}";
+                        $params .= "changeVal={$changeVal}";
                     }
-                    // echo "  act_zbConfigureReporting2 params = ${params}\n";
+                    // echo "  act_zbConfigureReporting2 params = {$params}\n";
                     $cmd['params'] = $params;
                     $commands2[$cmdJName] = $cmd;
                     $devUpdated = true;
-                    echo "  Cmd '${cmdJName}' UPDATED.\n";
+                    echo "  Cmd '{$cmdJName}' UPDATED.\n";
                 }
 
                 // Cluster 0001 updates
@@ -338,14 +338,14 @@
                     $cmd['params'] =  "clustId=0001&attrId=0021&attrType=20&minInterval=1800&maxInterval=3600";
                     $commands2[$cmdJName] = $cmd;
                     $devUpdated = true;
-                    echo "  Cmd '${cmdJName}' UPDATED.\n";
+                    echo "  Cmd '{$cmdJName}' UPDATED.\n";
                 } else if (($cmdFName == "act_zbConfigureReporting") &&
                     ($cmd['params'] ==  "ep=01&clustId=0001&attrType=20&attrId=0021&minInterval=0708&maxInterval=0E10&changeVal=")) {
                     $cmd['use'] = "act_zbConfigureReporting2";
                     $cmd['params'] =  "ep=01&clustId=0001&attrId=0021&attrType=20&minInterval=1800&maxInterval=3600";
                     $commands2[$cmdJName] = $cmd;
                     $devUpdated = true;
-                    echo "  Cmd '${cmdJName}' UPDATED.\n";
+                    echo "  Cmd '{$cmdJName}' UPDATED.\n";
                 } else if (($cmdFName == "BindToPowerConfig") && $oldSyntax) {
                     $cmdArr = Array(
                         "use"=> "zbBindToZigate",
@@ -414,35 +414,35 @@
                     $cmd['params'] =  "clustId=0006&attrType=10&attrId=0000";
                     $commands2[$cmdJName] = $cmd;
                     $devUpdated = true;
-                    echo "  Cmd '${cmdJName}' UPDATED.\n";
+                    echo "  Cmd '{$cmdJName}' UPDATED.\n";
                 } else if (($cmdFName == "act_zbConfigureReporting") &&
                     ($cmd['params'] ==  "ep=01&clustId=0006&attrType=10&attrId=0000&minInterval=0000&maxInterval=0000&changeVal=")) {
                     $cmd['use'] = "act_zbConfigureReporting2";
                     $cmd['params'] =  "ep=01&clustId=0006&attrId=0000&attrType=10";
                     $commands2[$cmdJName] = $cmd;
                     $devUpdated = true;
-                    echo "  Cmd '${cmdJName}' UPDATED.\n";
+                    echo "  Cmd '{$cmdJName}' UPDATED.\n";
                 } else if (($cmdFName == "act_zbConfigureReporting") &&
                     ($cmd['params'] ==  "ep=02&clustId=0006&attrType=10&attrId=0000&minInterval=0000&maxInterval=0000&changeVal=")) {
                     $cmd['use'] = "act_zbConfigureReporting2";
                     $cmd['params'] =  "ep=02&clustId=0006&attrId=0000&attrType=10";
                     $commands2[$cmdJName] = $cmd;
                     $devUpdated = true;
-                    echo "  Cmd '${cmdJName}' UPDATED.\n";
+                    echo "  Cmd '{$cmdJName}' UPDATED.\n";
                 } else if (($cmdFName == "act_zbConfigureReporting") &&
                     ($cmd['params'] ==  "ep=03&clustId=0006&attrType=10&attrId=0000&minInterval=0000&maxInterval=0000&changeVal=")) {
                     $cmd['use'] = "act_zbConfigureReporting2";
                     $cmd['params'] =  "ep=03&clustId=0006&attrId=0000&attrType=10";
                     $commands2[$cmdJName] = $cmd;
                     $devUpdated = true;
-                    echo "  Cmd '${cmdJName}' UPDATED.\n";
+                    echo "  Cmd '{$cmdJName}' UPDATED.\n";
                 } else if (($cmdFName == "act_zbConfigureReporting") &&
                     ($cmd['params'] ==  "ep=04&clustId=0006&attrType=10&attrId=0000&minInterval=0000&maxInterval=0000&changeVal=")) {
                     $cmd['use'] = "act_zbConfigureReporting2";
                     $cmd['params'] =  "ep=04&clustId=0006&attrId=0000&attrType=10";
                     $commands2[$cmdJName] = $cmd;
                     $devUpdated = true;
-                    echo "  Cmd '${cmdJName}' UPDATED.\n";
+                    echo "  Cmd '{$cmdJName}' UPDATED.\n";
                 } else if (($cmdFName == "etatDoor") && $oldSyntax) {
                     $cmdArr = Array(
                         "use" => "zb-0006-OnOff",
@@ -477,14 +477,14 @@
                     $cmd['params'] =  "clustId=0008&attrId=0000&attrType=20";
                     $commands2[$cmdJName] = $cmd;
                     $devUpdated = true;
-                    echo "  Cmd '${cmdJName}' UPDATED.\n";
+                    echo "  Cmd '{$cmdJName}' UPDATED.\n";
                 } else if (($cmdFName == "act_zbConfigureReporting") &&
                     ($cmd['params'] ==  "clustId=0008&attrType=10&attrId=0000&minInterval=0000&maxInterval=0000&changeVal=")) {
                     $cmd['use'] = "act_zbConfigureReporting2";
                     $cmd['params'] =  "clustId=0008&attrId=0000&attrType=20";
                     $commands2[$cmdJName] = $cmd;
                     $devUpdated = true;
-                    echo "  Cmd '${cmdJName}' UPDATED.\n";
+                    echo "  Cmd '{$cmdJName}' UPDATED.\n";
                 }
 
                 // Cluster 0102 updates
@@ -494,28 +494,28 @@
                     $cmd['params'] =  "clustId=0102&attrId=0008&attrType=20";
                     $commands2[$cmdJName] = $cmd;
                     $devUpdated = true;
-                    echo "  Cmd '${cmdJName}' UPDATED.\n";
+                    echo "  Cmd '{$cmdJName}' UPDATED.\n";
                 } else if (($cmdFName == "act_zbConfigureReporting") &&
                     ($cmd['params'] ==  "ep=01&clustId=0102&attrType=10&attrId=0000&minInterval=0000&maxInterval=0000&changeVal=")) {
                     $cmd['use'] = "act_zbConfigureReporting2";
                     $cmd['params'] =  "ep=01&clustId=0102&attrId=0008&attrType=20";
                     $commands2[$cmdJName] = $cmd;
                     $devUpdated = true;
-                    echo "  Cmd '${cmdJName}' UPDATED.\n";
+                    echo "  Cmd '{$cmdJName}' UPDATED.\n";
                 } else if (($cmdFName == "act_zbConfigureReporting") &&
                     ($cmd['params'] ==  "ep=01&clustId=0102&attrType=20&attrId=0000&minInterval=0000&maxInterval=0000&changeVal=")) {
                     $cmd['use'] = "act_zbConfigureReporting2";
                     $cmd['params'] =  "ep=01&clustId=0102&attrId=0008&attrType=20";
                     $commands2[$cmdJName] = $cmd;
                     $devUpdated = true;
-                    echo "  Cmd '${cmdJName}' UPDATED.\n";
+                    echo "  Cmd '{$cmdJName}' UPDATED.\n";
                 } else if (($cmdFName == "act_zbConfigureReporting") &&
                     ($cmd['params'] ==  "clustId=0102&attrType=20&attrId=0008&minInterval=0000&maxInterval=0000&changeVal=01")) {
                     $cmd['use'] = "act_zbConfigureReporting2";
                     $cmd['params'] =  "clustId=0102&attrId=0008&attrType=20&minInterval=0000&maxInterval=0000&changeVal=01";
                     $commands2[$cmdJName] = $cmd;
                     $devUpdated = true;
-                    echo "  Cmd '${cmdJName}' UPDATED.\n";
+                    echo "  Cmd '{$cmdJName}' UPDATED.\n";
                 } else if (($cmdFName == "WindowsCoveringStop") && $oldSyntax) {
                     $cmdArr = Array(
                         "use"=> "zbCmd-0102-Stop",
@@ -620,6 +620,15 @@
                     );
                     $devUpdated = true;
                     echo "  Cmd '".$cmdFName."' UPDATED.\n";
+                } else if ($cmdFName == "setReportLightColorSpectre") {
+                    $commands2["SetReporting 0300-0007"] = Array(
+                        "use" => "act_zbConfigureReporting2",
+                        "params" => "ep=#EP#&clustId=0300&attrId=0007&attrType=21",
+                        "execAtCreation" => "Yes",
+                        "execAtCreationDelay" => 11
+                    );
+                    $devUpdated = true;
+                    echo "  Cmd '".$cmdFName."' UPDATED.\n";
                 }
 
                 // Cluster 0405 updates
@@ -629,14 +638,14 @@
                     $cmd['params'] =  "ep=02&clustId=0405&attrId=0000&attrType=21&minInterval=300&maxInterval=600";
                     $commands2[$cmdJName] = $cmd;
                     $devUpdated = true;
-                    echo "  Cmd '${cmdJName}' UPDATED.\n";
+                    echo "  Cmd '{$cmdJName}' UPDATED.\n";
                 } else if (($cmdFName == "act_zbConfigureReporting") &&
                         ($cmd['params'] ==  "clustId=0405&attrType=21&attrId=0000&minInterval=012C&maxInterval=0258&changeVal=")) {
                     $cmd['use'] = "act_zbConfigureReporting2";
                     $cmd['params'] =  "clustId=0405&attrId=0000&attrType=21&minInterval=300&maxInterval=600";
                     $commands2[$cmdJName] = $cmd;
                     $devUpdated = true;
-                    echo "  Cmd '${cmdJName}' UPDATED.\n";
+                    echo "  Cmd '{$cmdJName}' UPDATED.\n";
                 }
 
                 // Cluster 0500 updates
@@ -646,7 +655,7 @@
                     $cmd['params'] =  "clustId=0500&attrId=0002&attrType=19";
                     $commands2[$cmdJName] = $cmd;
                     $devUpdated = true;
-                    echo "  Cmd '${cmdJName}' UPDATED.\n";
+                    echo "  Cmd '{$cmdJName}' UPDATED.\n";
                 } else if (($cmdFName == "BindShortToSmokeHeiman") && $oldSyntax) {
                     $cmdArr = Array(
                         "use"=> "zbBindToZigate",
@@ -666,7 +675,7 @@
                     $cmd['params'] =  "clustId=0B04&attrId=0508&attrType=21";
                     $commands2[$cmdJName] = $cmd;
                     $devUpdated = true;
-                    echo "  Cmd '${cmdJName}' UPDATED.\n";
+                    echo "  Cmd '{$cmdJName}' UPDATED.\n";
                 } else if (($cmdFName == "getPlugA") && $oldSyntax) {
                     $cmdArr = Array(
                         "use"=> "poll-0B04-0508",
@@ -686,6 +695,15 @@
                         "use"=> "poll-0B04-0505",
                     );
                     $commands2["Poll 0B04-0505"] = $cmdArr;
+                    $devUpdated = true;
+                    echo "  Cmd '".$cmdFName."' UPDATED.\n";
+                } else if ($cmdFName == "setReportPuissanceLegrand") {
+                    $commands2["SetReporting 0B04-050B"] = Array(
+                        "use" => "act_zbConfigureReporting2",
+                        "params" => "ep=#EP#&clustId=0B04&attrId=050B&attrType=29",
+                        "execAtCreation" => "Yes",
+                        "execAtCreationDelay" => 11
+                    );
                     $devUpdated = true;
                     echo "  Cmd '".$cmdFName."' UPDATED.\n";
                 }