+++ docs/fr_FR/Changelog.md
@@ -1,6 +1,8 @@
 # ChangeLog
 
 -   Correction: Interne: PHP8 warning.
+-   Améliorations: Modéles: Suppression cmd obsolete 'setReportLightColorSpectre' => 'act_zbConfigureReporting2'
+-   Améliorations: Modéles: Suppression cmd obsolete 'setReportPuissanceLegrand' => 'act_zbConfigureReporting2'
 
 ## 240920-BETA-3